+++ tests/Rules/Fixture/compiled/clean.php
@@ -1,0 +1,5 @@
+<?php
+// @bladestan-source: /project/resources/views/welcome.blade.php
+
+/** @var string $title */
+echo $title;